Apply more feedback (split up tests, whitespaces, call DecryptCekCore in DecryptCek, better resource name and value)
