+++ src/System.Security.Cryptography.Pkcs/src/Resources/Strings.resx
@@ -214,8 +214,8 @@
   <data name="Cryptography_Cms_Signing_RequiresPrivateKey" xml:space="preserve">
     <value>A certificate with a private key is required.</value>
   </data>
-  <data name="Cryptography_Cms_Decrypting_RequiresRSAPrivateKey" xml:space="preserve">
-    <value>RSA private key is required.</value>
+  <data name="Cryptography_Cms_Ktri_RSARequired" xml:space="preserve">
+    <value>An RSA key is required to decrypt for a RecipientInfo with a KeyTransport recipient type.</value>
   </data>
   <data name="Cryptography_Cms_TrustFailure" xml:space="preserve">
     <value>Certificate trust could not be established. The first reported error is: {0}</value>
