+++ src/System.Security.Cryptography.Pkcs/tests/EnvelopedCms/DecryptTests.cs
@@ -558,96 +558,5 @@ namespace System.Security.Cryptography.Pkcs.EnvelopedCmsTests.Tests
                 Assert.Equal<byte>(expectedContent.Content, contentInfo.Content);
             }
         }
-
-        public class DecryptTestsUsingCertWithPrivateKey : DecryptTests
-        {
-            public DecryptTestsUsingCertWithPrivateKey() : base(false) { }
-
-            [ConditionalFact(nameof(SupportsCngCertificates))]
-            [OuterLoop(/* Leaks key on disk if interrupted */)]
-            public static void DecryptMultipleRecipients()
-            {
-                // Force Decrypt() to try multiple recipients. Ensure that a failure to find a matching cert in one doesn't cause it to quit early.
-
-                CertLoader[] certLoaders = new CertLoader[]
-                {
-                    Certificates.RSAKeyTransfer1,
-                    Certificates.RSAKeyTransfer2,
-                    Certificates.RSAKeyTransfer3,
-                };
-
-                byte[] content = { 6, 3, 128, 33, 44 };
-                EnvelopedCms ecms = new EnvelopedCms(new ContentInfo(content), new AlgorithmIdentifier(new Oid(Oids.Aes256)));
-                CmsRecipientCollection recipients = new CmsRecipientCollection();
-                foreach (CertLoader certLoader in certLoaders)
-                {
-                    recipients.Add(new CmsRecipient(certLoader.GetCertificate()));
-                }
-                ecms.Encrypt(recipients);
-                byte[] encodedMessage = ecms.Encode();
-
-                ecms = new EnvelopedCms();
-                ecms.Decode(encodedMessage);
-
-                // How do we know that Decrypt() tries receipients in the order they appear in ecms.RecipientInfos? Because we wrote the implementation.
-                // Not that some future implementation can't ever change it but it's the best guess we have.
-                RecipientInfo me = ecms.RecipientInfos[2];
-
-                CertLoader matchingCertLoader = null;
-                for (int index = 0; index < recipients.Count; index++)
-                {
-                    if (recipients[index].Certificate.Issuer == ((X509IssuerSerial)(me.RecipientIdentifier.Value)).IssuerName)
-                    {
-                        matchingCertLoader = certLoaders[index];
-                        break;
-                    }
-                }
-                Assert.NotNull(matchingCertLoader);
-
-                using (X509Certificate2 cert = matchingCertLoader.TryGetCertificateWithPrivateKey())
-                {
-                    if (cert == null)
-                        return; // Sorry - CertLoader is not configured to load certs with private keys - we've tested as much as we can.
-                    X509Certificate2Collection extraStore = new X509Certificate2Collection();
-                    extraStore.Add(cert);
-                    ecms.Decrypt(extraStore);
-                }
-
-                ContentInfo contentInfo = ecms.ContentInfo;
-                Assert.Equal<byte>(content, contentInfo.Content);
-            }
-        }
-
-#if netcoreapp // API not present on netfx
-        public class DecryptTestsUsingExplicitPrivateKey : DecryptTests
-        {
-            public DecryptTestsUsingExplicitPrivateKey() : base(true) { }
-
-            [Fact]
-            public static void DecryptUsingWrongPrivateKeyType()
-            {
-                byte[] content = new byte[] { 1, 2, 3, 4 };
-                ContentInfo contentInfo = new ContentInfo(content);
-                AlgorithmIdentifier alg = new AlgorithmIdentifier(new Oid(Oids.Aes192));
-
-                EnvelopedCms ecms = new EnvelopedCms(contentInfo, alg);
-
-                using (X509Certificate2 cert = Certificates.RSAKeyTransfer1.GetCertificate())
-                {
-                    CmsRecipient cmsRecipient = new CmsRecipient(SubjectIdentifierType.IssuerAndSerialNumber, cert);
-                    ecms.Encrypt(cmsRecipient);
-                }
-
-                byte[] encoded = ecms.Encode();
-                ecms = new EnvelopedCms();
-                ecms.Decode(encoded);
-
-                using (ECDsa ecdsa = ECDsa.Create())
-                {
-                    Assert.Throws<CryptographicException>(() => ecms.Decrypt(ecms.RecipientInfos[0], ecdsa));
-                }
-            }
-        }
-#endif
     }
 }
